feat(frontend): add Schemas section; split Hook/Code controls
- New Schemas section (parameters/inputs/outputs JSON-schema editors) shown for
  both hook and code workflows, sourced from data.schemas. Section and each
  field collapsed by default; tool-card chrome (copy/collapse/line numbers).
  Writes route under data.schemas via {__siblingData:{schemas:...}}.
- Split HookCodeConfigControl into dedicated HookConfigControl and
  CodeConfigControl (one component per kind), consistent with SchemasConfigControl.
