+++ web/packages/agenta-entity-ui/src/DrillInView/SchemaControls/index.ts
@@ -26,8 +26,14 @@ export type {TextInputControlProps} from "./TextInputControl"
 export {EnumSelectControl} from "./EnumSelectControl"
 export type {EnumSelectControlProps} from "./EnumSelectControl"
 
-export {HookCodeConfigControl} from "./HookCodeConfigControl"
-export type {HookCodeConfigControlProps} from "./HookCodeConfigControl"
+export {HookConfigControl} from "./HookConfigControl"
+export type {HookConfigControlProps} from "./HookConfigControl"
+
+export {CodeConfigControl} from "./CodeConfigControl"
+export type {CodeConfigControlProps} from "./CodeConfigControl"
+
+export {SchemasConfigControl} from "./SchemasConfigControl"
+export type {SchemasConfigControlProps} from "./SchemasConfigControl"
 
 // ============================================================================
 // CONTROLS WITH CONTEXT INJECTION
